# docs/openapi.yaml
openapi: 3.0.3
info:
  title: FarmFlow API
  description: |
    Egyptian livestock marketplace API.
    All monetary values are in EGP (Egyptian Pound).
    All dates are ISO 8601 strings.
    Image paths are relative to the server root (e.g. `/uploads/filename.jpg`).

    **Authentication:** Pass a JWT token in the `Authorization: Bearer <token>` header.

    **Roles:**
    - `buyer` — can browse, order, review, manage favorites
    - `seller` — can manage listings, animals, dairy, supplies, finances
    - `admin` — full access including approval workflows

    **⚠️ Mobile app endpoint notes:**
    - `GET /api/listings/my` is a dedicated route (seller/admin JWT required) — returns own listings (all statuses)
    - `GET /api/orders/my` is handled by `GET /api/orders` (buyer JWT → own orders only)
    - `PATCH /api/listings/:id/status` does not exist — use `PUT /api/listings/:id` with a `status` body field
    - `PATCH /api/admin/users/:id` does not exist — use `PATCH /api/admin/users/:id/toggle`
  version: 1.0.0
  contact:
    email: a.khalil@soar.inc
servers:
  - url: http://localhost:5001/api
    description: Local development
  - url: https://api.farmflow.app/api
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MongoId:
      type: string
      pattern: '^[a-fA-F0-9]{24}$'
      example: 64f1a2b3c4d5e6f7a8b9c0d1
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: Unauthorized
    ValidationErrorResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            type: object
            properties:
              msg:   { type: string }
              param: { type: string }
              value: {}

    # ── Auth ────────────────────────────────────────────────────────────────────
    User:
      type: object
      properties:
        _id:           { $ref: '#/components/schemas/MongoId' }
        name:          { type: string }
        email:         { type: string, format: email }
        phone:         { type: string }
        role:          { type: string, enum: [buyer, seller, admin] }
        isActive:      { type: boolean }
        # Seller-only
        farmName:      { type: string }
        farmPhone:     { type: string }
        personalPhone: { type: string }
        governorate:   { type: string }
        farmDescription: { type: string }
        farmBanner:    { type: string, description: '/uploads/...' }
        farmLocation:
          type: object
          properties:
            lat: { type: number }
            lng: { type: number }
            address: { type: string }
        farmCertificates:
          type: array
          items: { type: string }
        averageRating:  { type: number }
        reviewCount:    { type: integer }
        fcmToken:       { type: string, nullable: true, description: Firebase Cloud Messaging device token (set by Flutter app) }
        createdAt:      { type: string, format: date-time }
    AuthResponse:
      type: object
      properties:
        token: { type: string }
        user:  { $ref: '#/components/schemas/User' }

    # ── Listings ────────────────────────────────────────────────────────────────
...
    Budget:
      type: object
      properties:
        _id:      { $ref: '#/components/schemas/MongoId' }
        seller:   { $ref: '#/components/schemas/MongoId' }
        year:     { type: integer }
        month:    { type: integer, minimum: 1, maximum: 12 }
        category: { type: string }
        amount:   { type: number }

    # ── AppConfig (Eid) ──────────────────────────────────────────────────────────
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, password, role]
              properties:
                role:          { type: string, enum: [buyer, seller] }
                name:          { type: string }
                email:         { type: string, format: email, description: Required for buyer }
                phone:         { type: string }
                password:      { type: string, minLength: 6 }
                nationalId:    { type: string, description: 14-digit Egyptian national ID }
                # Seller-only
                farmName:      { type: string }
                farmPhone:     { type: string }
                personalPhone: { type: string }
                governorate:   { type: string }
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '400':
          description: Validation error or duplicate field
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ValidationErrorResponse' }
        '409':
          description: Email / phone / national ID already registered
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /auth/login:
    post:
      tags: [Auth]
      summary: Login with email/phone and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier, password]
              properties:
                identifier: { type: string, description: Email or phone number }
                password:   { type: string }
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /auth/me:
    get:
      tags: [Auth]
      summary: Get authenticated user profile
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401':
          description: Token missing or invalid
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /auth/profile:
    put:
      tags: [Auth]
      summary: Update user profile (supports multipart for farmBanner image)
      security: [{ BearerAuth: [] }]
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                name:            { type: string }
                phone:           { type: string }
                governorate:     { type: string }
                farmName:        { type: string }
                farmPhone:       { type: string }
                personalPhone:   { type: string }
                farmDescription: { type: string }
                farmBanner:      { type: string, format: binary }
                farmLocation:    { type: string, description: 'JSON string of {lat, lng, address}' }
                farmCertificates: { type: array, items: { type: string } }
          application/json:
            schema:
              type: object
              properties:
                name:          { type: string }
                phone:         { type: string }
                governorate:   { type: string }
                farmName:      { type: string }
                farmPhone:     { type: string }
                personalPhone: { type: string }
                farmDescription: { type: string }
                farmLocation:
                  type: object
                  properties:
                    lat: { type: number }
                    lng: { type: number }
                    address: { type: string }
                farmCertificates: { type: array, items: { type: string } }
      responses:
        '200':
          description: Updated user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { description: Unauthorized }
        '409':
          description: Phone / email already used
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /auth/password:
    put:
      tags: [Auth]
      summary: Change password (requires current password verification)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword: { type: string }
                newPassword:     { type: string, minLength: 6 }
      responses:
        '200':
          description: Password changed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string, example: Password updated successfully }
        '400':
          description: Current password incorrect or validation failed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /auth/fcm-token:
    patch:
      tags: [Auth]
      summary: Register or refresh Firebase Cloud Messaging token (Flutter app)
      description: Called by the Flutter app after FCM permission granted. Saves the device token to the authenticated user's profile for server-side push notifications.
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string, description: FCM device token from FirebaseMessaging.instance.getToken() }
      responses:
        '200':
          description: Token saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, example: true }
        '400':
          description: Token field missing
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /auth/verify-id:
    post:
      tags: [Auth]
      summary: Validate Egyptian national ID format and extract info
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nationalId]
              properties:
                nationalId: { type: string, description: 14-digit national ID }
      responses:
        '200':
          description: ID is structurally valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  valid:       { type: boolean }
                  birthDate:   { type: string, format: date }
                  age:         { type: integer }
                  governorate: { type: string }
        '400':
          description: Invalid format
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  # ── SELLERS ──────────────────────────────────────────────────────────────────
...
  /budget:
    get:
      tags: [Budget]
      summary: Get budget entries for a year (seller only)
      security: [{ BearerAuth: [] }]
      parameters:
        - in: query
          name: year
          schema: { type: integer, example: 2025 }
      responses:
        '200':
          description: Array of budget entries
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Budget' }

    put:
      tags: [Budget]
      summary: Upsert budget target (seller only)
      description: Creates or updates a budget entry for a specific year/month/category combination.
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [year, month, category, amount]
              properties:
                year:     { type: integer }
                month:    { type: integer, minimum: 1, maximum: 12 }
                category: { type: string }
                amount:   { type: number, minimum: 0 }
      responses:
        '200':
          description: Upserted budget entry
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Budget' }